[Windows only] -remove code to turn off screensaver due to problems -updated version number to b33 -added icon for add/remove programs list [All platforms] -for the flock size in the F2 display, if playing gold sheep then report gold flock size -shorten delay when no sheep are available -show native string for open file in log

+++ trunk/client_generic/ContentDownloader/Shepherd.cpp
@@ -57,8 +57,8 @@ using namespace std;
 //
 uint64 Shepherd::s_ClientFlockBytes = 0;
 uint64 Shepherd::s_ClientFlockGoldBytes = 0;
-size_t Shepherd::s_ClientFlockCount = 0;
-size_t Shepherd::s_ClientFlockGoldCount = 0;
+uint64 Shepherd::s_ClientFlockCount = 0;
+uint64 Shepherd::s_ClientFlockGoldCount = 0;
 char *Shepherd::fRootPath = NULL;
 char *Shepherd::fMpegPath = NULL;
 char *Shepherd::fXmlPath = NULL;
